refactor #260512: App, BoardPage で、SRP 分割
