+++ .github/reports/issue-49-53-check.md
@@ -1,0 +1,23 @@
+# Issue #49-#53 検証レポート
+
+作成日時: 2026-05-08T17:10:18+09:00
+作成者: Copilot CLI
+
+概要:
+- 変更を確認した結果、ダッシュボードと認証ストアのリファクタリングは追加されているが、いくつか未解決の点が残っている。
+
+確認したファイル:
+- src/store/useAuthStore.ts — 認証ストアの自動選択ロジック削除、deselectProfile の追加、createProfile の実装あり
+- src/pages/DashboardPage.tsx — ダッシュボード一覧と新規作成モーダル呼び出しを実装
+- src/App.tsx — /dashboard ルート追加、ヘッダーにボード一覧リンク、ログイン後の遷移制御を追加
+- src/components/board/BoardPage.tsx — ログインガードと簡潔化されたレンダリング（ただし未解決点あり）
+- src/components/common/CreateProfileModal.tsx — プロファイル作成 UI（自動でボードを開く処理は未追加）
+
+未解決（推奨対応）:
+1. CreateProfileModal: 新規ボード作成後に自動でボードを開く（navigate('/') を呼ぶ）。
+2. BoardPage: プロファイル未選択で直接アクセスされた場合はスピナーではなく /dashboard へリダイレクトする安全ガードを追加する。
+
+受け渡しメモ:
+- ユーザー指定により自動修正は行っていません。修正を希望する場合は対応可。
+
+以上。
