+++ src/hooks/useAuthGuard.ts
@@ -1,0 +1,10 @@
+import { useEffect } from 'react';
+import { useAuthStore } from '../store/useAuthStore';
+
+export const useAuthGuard = () => {
+  const checkAuth = useAuthStore((state) => state.checkAuth);
+
+  useEffect(() => {
+    checkAuth();
+  }, [checkAuth]);
+}
